refactor : SellNowService 내부 로직 중 의존에 따른 역할분리
